Fixed bug in code.
This code creates a commit message of 50 characters or less by using the OpenAI library to generate a summary of a diff, then using the QAStuffChain to generate a subject from the summary and the question. The postProcess function is used to trim the subject and body of the commit message.