Installed git-magic-commit package with MIT License.
This project provides instructions on how to install and use the 'git-magic-commit' package, which is licensed under the MIT License. It requires the user to install the package globally using npm and set up their OpenAI API credentials as environment variables. Once installed, the user can use the command 'git-magic-commit' to use the package.