Add new function to create commit message subject.
This code is making changes to the create-commit-message.ts and create-commit-message2.ts files, as well as the git-magic-commit.ts and lib/git.ts files. The changes include importing new libraries, updating types, and making changes to the post-processing of commit messages. The code is also adding a new function to create a subject for the commit message.